BUG: Fix 64-to-32 bit truncations, including a real overflow in fslio
31 -Wshorten-64-to-32 warnings.  Most were benign; one was not.

FslSeekVolume() declared `int offset` and assigned it a byte position
into the image file:

    offset = fslio->niftiptr->iname_offset
             + vols * FslGetVolSize(fslio) * fslio->niftiptr->nbyper;
    return znzseek(fslio->fileptr, offset, SEEK_SET);

Seek to any volume past 2GB and the multiplication overflows, so znzseek
-- which takes a 64-bit znz_off_t -- receives a wrong and quite possibly
negative position.  offset is a local, so widening it to znz_off_t and
widening the operands changes no interface.  FslReadVolumes' volbytes had
the same shape and is now size_t.

The rest fall into two groups.  Internal helpers were widened to carry
the value they were already being handed: nifti_read_extensions() and
rci_alloc_mem() now return int64_t rather than truncating their own
results, and nifti_read_next_extension() and nifti_check_extension() take
an int64_t `remain`.  nt_read_bricks(), declared in the uninstalled
nifti_tool.h, takes an int64_t length.

Where the narrowing target is an installed prototype it can only be made
explicit, so each such cast carries a note: nifti_image_load_bricks()
returning a brick count as int, nifti_read_subregion_image() returning a
byte count as int, nifti_read_ascii_image() taking an int header length,
and FslSeekVolume()'s int return.  The remainder are strlen() results
assigned to int and fread()/znzread() counts, bounded by buffers a few
hundred bytes long.

Two format strings were widened to PRId64 to match the parameters that
changed.

+++ cifti/afni_xml.c
@@ -211,7 +211,7 @@ afni_xml_list axml_read_file(const char * fname, int read_data)
    {
       if( reset_xml_buf(xd, &buf, &bsize) ) break;
 
-      blen = fread(buf, 1, bsize, fp);
+      blen = (unsigned)fread(buf, 1, bsize, fp);
       if( ferror(fp) ) {
          fprintf(stderr,"** failed to read XML file '%s'\n", fname);
          break;
@@ -307,7 +307,7 @@ afni_xml_list axml_read_buf(const char * buf_in, int64_t bin_len)
 
         /* decide how much to copy and copy */
         if( bin_remain >= bsize ) blen = bsize;
-        else                      blen = bin_remain;
+        else                      blen = (unsigned)bin_remain;
 
         if(blen > 0 && blen <= (unsigned)bsize) {
            memcpy(buf, bin_ptr, blen);
@@ -891,7 +891,7 @@ static const char * strip_whitespace(const char * str, int slen)
    /* if string is long, forget it */
    if( !str || slen > 1024 ) return str;
 
-   len = strlen(str);
+   len = (int)strlen(str);
    if( slen > 0 && slen < len ) len = slen;
    if( len <= 0 ) return str;
 
@@ -985,7 +985,7 @@ static int disp_gen_text(afni_xml_control * xd, const char * header,
 
    if( len == wlen ) {
        str = "whitespace";     /* just note the whitespace */
-       len = strlen(str);
+       len = (int)strlen(str);
    }
 
    show_depth(xd, 1);
